Add noop implementation
